requirements.txt
langchain
langsmith
langgraph
ipykernel
python-dotenv
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence-transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-openai
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve[all]
duckduckgo-search
validators
youtube-transcript-api
unstructured
pytube
numexpr
huggingface_hub
cassio
datasets
tiktoken
PyPDF2
openai
gradio
langchain_nvidia_ai_endpoints
pinecone-client
pinecone-text
pinecone-notebooks
neo4j == 5.28.1
langchain-neo4j
fitz
frontend
tools
faker
